refactor : CategoryService 및 repository 구조 변경
